% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{connectionStringAccess}
\alias{connectionStringAccess}
\title{Connection String Access}
\usage{
connectionStringAccess(mdb, uid = "", pwd = "",
  globalPartialBulkOps = 0)
}
\arguments{
\item{mdb}{full path to MS Access file}

\item{uid}{user id, if any}

\item{pwd}{password, if any}

\item{globalPartialBulkOps}{A Long value (read/write) that determines the
behavior of the Jet database engine when SQL DML bulk operations fail. When
set to allow partial completion of bulk operations, inconsistent changes
can occur, because operations on some records could succeed and others
could fail. When set to allow no partial completion of bulk operations, all
changes are rolled back if a single error occurs. The Jet OLEDB:Global
Partial Bulk Ops property setting can be overridden for the current
Recordset object by setting the Jet OLEDB:Partial Bulk Ops property.The Jet
OLEDB:Global Partial Bulk Ops and Jet OLEDB:Partial Bulk Ops properties can
be set to any of the following values: Default = 0, Partial = 1, No Partial
= 2}
}
\description{
Connection String Access
}
\references{
\url{http://msdn.microsoft.com/en-us/library/office/aa140022\%28v=office.10\%29.aspx}
}
