% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_R_interface.R
\name{createRDatabaseInterface}
\alias{createRDatabaseInterface}
\title{Create R-Function Interface to Database}
\usage{
createRDatabaseInterface(
  db = NULL,
  dbSchema = NULL,
  tableNamePattern = "",
  functionPrefix = "db_",
  pattern = "^",
  rfile = file.path(tempdir(), "dbInterface.R"),
  rfile.create = file.path(tempdir(), "dbInterfaceCreate.R"),
  create.create.functions = FALSE,
  dbg = TRUE,
  sql.dbg = TRUE
)
}
\arguments{
\item{db}{name of ODBC data source or full path to MS Access or Excel file}

\item{dbSchema}{list structure describing the database schema, as returned by
kwb.odm::dbSchema_ODM()}

\item{tableNamePattern}{pattern matching the names of tables/views for which
an accessor function is to be generated}

\item{functionPrefix}{prefix to be given to each generated function. Default:
"db_"}

\item{pattern}{pattern matching the part of the table name that will be
replaced with \emph{functionPrefix}. Default: "^", i.e. the prefix will be
appended to the start of the table name}

\item{rfile}{full path to the file in which to store the data access 
functions}

\item{rfile.create}{full path to the file in which to store the data creation 
functions (only relevant if \emph{crete.create.functions} is TRUE)}

\item{create.create.functions}{logical. If TRUE (the default if FALSE) not 
only data access but also data creation functions are generated}

\item{dbg}{if TRUE, progress messages are shown during the creation of the
interface functions}

\item{sql.dbg}{if TRUE, SQL strings used to access the database are shown
when calling the interface functions}
}
\description{
Create R-Function Interface to Database
}
