% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{currentDb}
\alias{currentDb}
\title{Get Path to or Name of Current Database}
\usage{
currentDb(dbg = TRUE)
}
\arguments{
\item{dbg}{if TRUE, a message obout setting the current database is printed}
}
\description{
Get Path to or name of current database (as set with 
\code{\link{setCurrentDb}})
}
