% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dumpDatabase}
\alias{dumpDatabase}
\title{Export Database Tables to CSV Files}
\usage{
dumpDatabase(db, pattern = "^tbl", target_dir = NULL,
  create_target_dir = FALSE, sep = ",", dec = ".", qmethod = "double",
  row.names = FALSE, ...)
}
\arguments{
\item{db}{full path to database or name of ODBC data source}

\item{pattern}{pattern matching names of tables to be exported. Default:
"^tbl", i.e. tables starting with "tbl"}

\item{target_dir}{target directory. By default a new directory is created in
the same directory as mdb resides in. The new directory has the same name
as the database file with dots substituted with underscores}

\item{create_target_dir}{if \code{TRUE}, the target directory \code{tdir} is
created if it does not exist.}

\item{sep}{passed to \code{\link[utils]{write.table}}}

\item{dec}{passed to \code{\link[utils]{write.table}}}

\item{qmethod}{passed to \code{\link[utils]{write.table}}}

\item{row.names}{passed to \code{\link[utils]{write.table}}}

\item{\dots}{further arguments passed to \code{\link[utils]{write.table}}}
}
\description{
Exports all tables of a database of which the names match a given pattern to
csv files.
}
