% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current.R
\name{getCurrentSqlDialect}
\alias{getCurrentSqlDialect}
\title{Get Current SQL Dialect}
\usage{
getCurrentSqlDialect(warn = TRUE, dbg = FALSE)
}
\arguments{
\item{warn}{if TRUE and if no current SQL dialog is stored in the options,
the program stops with an error message}

\item{dbg}{if TRUE, a message about the current SQL dialect is printed}
}
\description{
Get Current SQL Dialect
}
