% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_R_interface.R
\name{getFilteredRecords}
\alias{getFilteredRecords}
\title{Get Filtered Records}
\usage{
getFilteredRecords(db, tableName, keyValues, fields, like, ...)
}
\arguments{
\item{db}{database name or database file}

\item{tableName}{name of the table from which to read records}

\item{keyValues}{list of \code{key = value} pairs with the keys representing
field names of the table \code{tableName}, defining filter criteria}

\item{fields}{table fields to be selected}

\item{like}{if \code{TRUE} the SQL \code{LIKE} operator is used instead of
the equal operator \code{=} when matching the values in \code{keyValues} 
with the table fields}

\item{\dots}{additonal arguments passed to \code{\link{selectFromDb}}}
}
\description{
Get Filtered Records
}
