% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNamedExcelRanges.R
\name{getNamedExcelRanges}
\alias{getNamedExcelRanges}
\title{Get Named Ranges from Excel File}
\usage{
getNamedExcelRanges(file, pattern = "^range", dbg = TRUE,
  stringsAsFactors = FALSE, ...)
}
\arguments{
\item{file}{path to Excel file}

\item{pattern}{pattern matching the names of named cell ranges to be read.
By default, all ranges starting with \code{range} are read.}

\item{dbg}{logical. If \code{TRUE} debug messages are shown}

\item{stringsAsFactors}{passed to \code{\link{hsGetTable}}}

\item{\dots}{further arguments passed to \code{\link{hsGetTable}}}
}
\value{
list of data frames each of which represents the content a named cell 
  range in the Excel \code{file} and each of which was read with 
  \code{\link{hsGetTable}}.
}
\description{
Get Named Ranges from Excel File
}
