% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getSqlDialect}
\alias{getSqlDialect}
\title{Get SQL Dialect from Given Database}
\usage{
getSqlDialect(db, use2007Driver = NULL)
}
\arguments{
\item{db}{ODBC database name or full path to database (mdb or xls)}

\item{use2007Driver}{passed to \code{\link{isMySQL}}}
}
\description{
get SQL dialect ("mysql" or "msaccess") from given database
}
