% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{getTypeIdentifier}
\alias{getTypeIdentifier}
\title{Get Type Identifier}
\usage{
getTypeIdentifier(x)
}
\arguments{
\item{x}{R object for which to find an adequate database object type}
}
\value{
Returns "int", "double", "text", "date_time" or "boolean" depending
  on the data type of \emph{x}
}
\description{
Get Type Identifier
}
