% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hsClearTable}
\alias{hsClearTable}
\title{Clear a Database Table}
\usage{
hsClearTable(mdb, tbl, cond = TRUE, ...)
}
\arguments{
\item{mdb}{full path to MS Access database (*.mdb).}

\item{tbl}{table name.}

\item{cond}{optional: condition.}

\item{\dots}{additional arguments passed to hsSqlQuery, e.g. "errors=TRUE"}
}
\description{
Deletes all the rows of the table \eqn{tbl}. This function performs opening
of the connection, clearing of the table and closing of the connection. If
an error occurs the program stops and an error description is shown.
}
\seealso{
\code{\link{hsDropTable}}
}
