% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{hsCloseDb}
\alias{hsCloseDb}
\title{Close Connection to MS Access or Excel}
\usage{
hsCloseDb(con)
}
\arguments{
\item{con}{Open database connection as returned by
\link{hsOpenDb}/\code{odbcConnect}}
}
\description{
Closes the database connection.
}
\seealso{
\code{\link{hsOpenDb}}
}
