% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{hsCloseMdb}
\alias{hsCloseMdb}
\title{Close Connection}
\usage{
hsCloseMdb(con)
}
\arguments{
\item{con}{Open database connection as returned by
\link{hsOpenMdb}/odbcConnect}
}
\description{
Deprecated. Use code{\link{hsCloseDb}} instead.
}
