% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hsDropTable}
\alias{hsDropTable}
\title{Drop Database Table(s)}
\usage{
hsDropTable(mdb, tbl, isPtrn = FALSE, dbg = TRUE)
}
\arguments{
\item{mdb}{full path to MS Access database (*.mdb).}

\item{tbl}{table name.}

\item{isPtrn}{if TRUE, \emph{tbl} is interpreted as a regular expression
matching the names of the tables to be deleted.}

\item{dbg}{if TRUE, debug messages are shown}
}
\description{
Removes the table \eqn{tbl} (if permitted). This function performs opening of
the connection, dropping of the table and closing of the connection.  If an
error occurs the program stops and an error description is shown.
}
\seealso{
\code{\link{hsClearTable}}
}
