% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsFields.R
\name{hsFields}
\alias{hsFields}
\title{Available Fields in Database Table}
\usage{
hsFields(
  mdb,
  tbl,
  namesOnly = TRUE,
  chopDollar = TRUE,
  ignore.case = (!isMySQL(mdb, use2007Driver = use2007Driver)),
  use2007Driver = NULL,
  dbg = FALSE
)
}
\arguments{
\item{mdb}{full path to MS Access database file (extension \dQuote{.mdb} or 
\dQuote{.accdb}) or MS Excel file (extension \dQuote{.xls} or 
\dQuote{.xlsx}).}

\item{tbl}{table name.}

\item{namesOnly}{if TRUE, only field names are returned, otherwise all
available information on the fields. Default: TRUE}

\item{chopDollar}{if TRUE (default), a dollar sign at the end of the table
name is removed before sending it to \code{sqlColumns},}

\item{ignore.case}{if TRUE, case is ignored when comparing the given table
with the names of the existing tables. Default: FALSE}

\item{use2007Driver}{passed to \code{\link{isMySQL}}}

\item{dbg}{if TRUE, debug messages are shown}
}
\value{
Vector containing the field names of the database table (if namesOnly
  = TRUE) or data.frame with columns \emph{COLUMN_NAME}, \emph{DATA_TYPE},
  \emph{TYPE_NAME}, \emph{COLUMN_SIZE}, \emph{BUFFER_LENGTH},
  \emph{DECIMAL_DIGITS}, \emph{NUM_PREC_RADIX}, \emph{NULLABLE} describing
  the database fields in detail, otherwise.
}
\description{
Returns a vector containing the field names of a database table.
}
\examples{
\dontrun{ 
## List the fields of table "tbl_Hyd" in the example database
## (only on Windows!)

if (.Platform$OS.type == "windows") {
  setCurrentSqlDialect("msaccess")
  fields <- hsFields(xmdb(), "tbl_Hyd")
  fields
}
  
## Ouput:
# [1] "Zeitst"  "Q"       "v"       "H"       "T_Kanal"
}

}
\seealso{
\code{\link{hsTables}}
}
