% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_specifics.R
\name{hsJetDate}
\alias{hsJetDate}
\title{Date to \dQuote{mm/dd/yyyy HH:MM:SS}}
\usage{
hsJetDate(datetime, dbg = FALSE)
}
\arguments{
\item{datetime}{Date (and time) information in forms of Date object or POSIX
object or string.}

\item{dbg}{if TRUE, debug messages are shown}
}
\description{
Returns a date in MS Jet SQL syntax: mm/dd/yyyy HH:MM:SS
}
