% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hsMdbTimeSeries}
\alias{hsMdbTimeSeries}
\title{Get Mdb time series in UTC}
\usage{
hsMdbTimeSeries(mdb, tbl, tsField = hsTsField(mdb, tbl), fields = "*",
  minDate = NULL, maxDate = NULL, resolution = "min", inclLast = TRUE,
  sqlFilter = "TRUE", dbg = FALSE, calcType = 1)
}
\arguments{
\item{mdb}{Full path to MS Access database file (*.mdb)}

\item{tbl}{Name of table containing the time-series data.}

\item{tsField}{Name of table field containing the timestamps.}

\item{fields}{Vector containing names of value fields to be selected from the
table. This vector may or may not contain the name of the timetamp field.}

\item{minDate}{Minimum date (and time) of time interval to be selected in
ISO-Syntax: yyyy-mm-dd [HH:MM:SS], where the part in brackets in optional.}

\item{maxDate}{Day following the maximum date of the time interval to be
selected, in ISO-Syntax: yyyy-mm-dd [HH:MM:SS], where the part in brackets
in optional.}

\item{resolution}{time resolution: \dQuote{min} = minutes, \dQuote{s} =
seconds. If time resolution is \dQuote{min} timestamps are rounded to the
next full minute.}

\item{inclLast}{If TRUE, \emph{maxDate} will be included in result data set,
else excluded.}

\item{sqlFilter}{additional SQL filter criterion}

\item{dbg}{if TRUE, debug messages are shown}

\item{calcType}{for internal use only, do not change!}
}
\value{
data.frame with POSIXct timestamp column <strTimestamp> (UTC time
  zone) and value columns as selected in <strFieldList>
}
\description{
Reads time-series data from an MS Access database table and returns a data
frame containing the data. The name of the timestamp field must be given in
\emph{tsField} and the names of the value fields to be selected from the
table must be given in vector \emph{fields}. Instead of an ODBC channel the
name of the database must be given. This function takes care that the
timestamps are transferred correctly between MS Access and R by requesting
date and time information separately from MS Access and putting both together
to a POSIXct object in UTC timezone. This is necessary because with very long
data sets the RODBC function sqlQuery (or the function
\code{\link{hsSqlQuery}} that calls this function) may deliver timestamps in
which time information is lacking!
}
\examples{
\dontrun{
## Get flow time series of 24 of August 2011 from tbl_Hyd in example database

if (.Platform$OS.type == "windows") {

  ts <- hsMdbTimeSeries(
    xmdb(), "tbl_Hyd", "Zeitst", c("Q", "v"), "2011-08-24", "2011-08-25"
  )
  
  ## Show the last records of the returned dataset.
  tail(ts)
}
   
## Output:
#                   Zeitst     Q     v
# 1435 2011-08-24 23:55:00 0.638 0.281
# 1436 2011-08-24 23:56:00 0.601 0.265
# 1437 2011-08-24 23:57:00 0.564 0.249
# 1438 2011-08-24 23:58:00 0.536 0.237
# 1439 2011-08-24 23:59:00 0.504 0.223
# 1440 2011-08-25 00:00:00 0.483 0.214
}

}
\seealso{
\code{\link{hsGetTimeSeries}, \link{hsGetTable}}
}
