% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{hsOpenDb}
\alias{hsOpenDb}
\title{Open Connection to MS Access or Excel}
\usage{
hsOpenDb(src, use2007Driver = NULL, dbg = FALSE, DBMSencoding = "", ...)
}
\arguments{
\item{src}{full path to MS Access database file (extension \dQuote{.mdb} or 
\dQuote{.accdb}) or MS Excel file (extension \dQuote{.xls} or 
\dQuote{.xlsx}) or name of ODBC data source.}

\item{use2007Driver}{if TRUE the functions odbcConnectAccess2007 and
odbcConnectExcel2007 are used instead of odbcConnectAccess and
odbcConnectExcel, respectively}

\item{dbg}{if TRUE and if the connection could be established details of the 
connection are shown.}

\item{DBMSencoding}{finally passed to \code{odbcDriverConnect}. Default: "",
You may want to use: "UTF-8"}

\item{\dots}{additional arguments passed to the odbcConnect...()-function}
}
\value{
On success an object of class RODBC describing the connection is
  returned. On failure nothing is returned and program execution stops.
}
\description{
Opens database connection to MS Access or MS Excel, checks the connection and
stops on error.
}
\examples{
\dontrun{ 
## Open a connection to the example database
## (only on Windows!)

if (.Platform$OS.type == "windows") {

  con <- hsOpenDb(xmdb())
  con
}
   
## Details on the established connection are shown:
# RODBC Connection 9
# Details:
#   case=nochange
#   DBQ=C:\\Users\\hsonne\\Documents\\R\\win-library\\2.14\\kwb.base\\...
#   Driver={Microsoft Access Driver (*.mdb)}
#   DriverId=25
#   FIL=MS Access
#   MaxBufferSize=2048
#   PageTimeout=5
#   UID=admin
  
## Close the connection again

if (.Platform$OS.type == "windows") {

  hsCloseDb(con)
}
}
   
}
\seealso{
\code{\link{hsCloseDb}}
}
