% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{hsOpenMdb}
\alias{hsOpenMdb}
\title{Open Connection to MS Access Database}
\usage{
hsOpenMdb(mdb, dbg = FALSE)
}
\arguments{
\item{mdb}{full path to MS Access database file.}

\item{dbg}{if TRUE and if the connection could be established details of the
connection are shown.}
}
\description{
Deprecated. Use \code{\link{hsOpenDb}} instead.
}
