% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{hsPutTable}
\alias{hsPutTable}
\title{Save Data Frame to Database Table}
\usage{
hsPutTable(mdb, myData, tbl = "tblTmp", types = NULL, overwrite = FALSE,
  DBMSencoding = "", dbg = TRUE)
}
\arguments{
\item{mdb}{full path to MS Access database file (*.mdb)}

\item{myData}{data.frame containing data to be written to database table}

\item{tbl}{Name of table to be created in the database}

\item{types}{field types to be passed to sqlSave as argument \emph{varTypes},
see ?sqlSave for details.}

\item{overwrite}{shall existing table be overwritten?}

\item{DBMSencoding}{finally passed to \code{odbcDriverConnect}. Default: "",
You may want to use: "UTF-8"}

\item{dbg}{if TRUE, debug messages are shown}
}
\value{
In case of success the name of the created table is returned.
}
\description{
Writes data to a database table. This function performs opening of the
connection, saving of the data to a table and closing of the connection.  If
an error occurs the program stops and an error description is shown. If a
table named \emph{tbl} already exists in the database \emph{mdb} the existing
table is only overwritten if \emph{overwrite} is TRUE. Otherwise a
\code{\link[kwb.utils]{hsSafeName}} will be found for the table.
}
\examples{
\dontrun{
## Create a data.frame
df1 <- data.frame(id = 1:4, rnd = sample(1:100)[1:4])
  
## Write data.frame into a table in the example database; as no
## table name is specified, a table name is generated. The name
## of the table is returned.
## (only on Windows!)

if (.Platform$OS.type == "windows") {

  tbl <- hsPutTable(xmdb(), df1)
  tbl # table name here: [1] "tblTmp"

   
  ## Get the data from the created table back again and print the
  ## data. As we see, a table containing four different random
  ## numbers between one and 100 has been created.
  df2 <- hsGetTable(xmdb(), tbl)
  df2
}
   
## Output:
#   id rnd
# 1  1  82
# 2  2  14
# 3  3  99
# 4  4   6
}

}
\seealso{
\code{\link{hsSqlQuery}, \link{hsGetTable}}
}
