% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hsSetForeignKey}
\alias{hsSetForeignKey}
\title{Set Foreign Key for Table Field}
\usage{
hsSetForeignKey(mdb, tbl, field, ref.tbl, ref.field,
  key.name = .getForeignKeyName(tbl, field, ref.tbl, ref.field),
  dbg = FALSE)
}
\arguments{
\item{mdb}{full path to MS Access database file (*.mdb)}

\item{tbl}{name of table containing the field for which the foreign key
constraint is to be defined}

\item{field}{name of table field for which the foreign key constraint is to
be defined}

\item{ref.tbl}{name of table containing the referenced foreign key field}

\item{ref.field}{name of foreign key field in ref.tbl}

\item{key.name}{optional. Name to be given to the foreign key}

\item{dbg}{passed to \code{\link{hsSqlQuery}}}
}
\description{
Set foreign key constraint for a table field
}
