% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hsSetPrimaryKey}
\alias{hsSetPrimaryKey}
\title{Set Primary Key of Database Table}
\usage{
hsSetPrimaryKey(mdb, tbl, keyFields, dbg = FALSE)
}
\arguments{
\item{mdb}{Full path to MS Access database file (*.mdb).}

\item{tbl}{Name of table in which key fields are to be defined.}

\item{keyFields}{(Vector of) key field name(s)}

\item{dbg}{if TRUE, debug messages are shown}
}
\description{
Sets fields with names given in vector \emph{keyFields} as key fields of 
table \emph{tbl} in MS ACCESS database \emph{mdb}.
}
