% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_expressions.R
\name{hsSqlExAnd}
\alias{hsSqlExAnd}
\title{SQL Expression \dQuote{AND}}
\usage{
hsSqlExAnd(x, bFunc)
}
\arguments{
\item{x}{vector of strings, e.g. representing table field names.}

\item{bFunc}{name of a boolean function to be \dQuote{applied} to each
element of \emph{x}.}
}
\description{
Creates a boolean expression of the form \cr \code{bFunc(x1) AND bFunc(x2)
AND ... AND bFunc(xn)}. This function can be used to create SQL queries where
many table fields have to be checked in the same way for some criterion (see
example).
}
\examples{
## Build SQL query finding records in table t in which all
## of the table fields f1 to f100 are NULL.
sql <- sprintf("SELECT * FROM t WHERE \%s", 
               hsSqlExAnd(paste("f", 1:100, sep = ""), "isNull"))
  
## Show the SQL string
sql
  
## Output (middle part omitted):
# SELECT * FROM t WHERE (TRUE) AND isNull(f1) AND 
# isNull(f2) AND isNull(f3) AND ... AND isNull(f100)

}
\seealso{
\code{\link{hsSqlExOr}}
}
