% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_expressions.R
\name{hsSqlExOr}
\alias{hsSqlExOr}
\title{SQL Expression \dQuote{OR}}
\usage{
hsSqlExOr(x, bFunc)
}
\arguments{
\item{x}{vector of strings, e.g. representing table field names.}

\item{bFunc}{name of a boolean function to be \dQuote{applied} to each
element of \emph{x}.}
}
\description{
Creates a boolean expression of the form \cr \code{bFunc(x1) OR bFunc(x2) OR 
... OR bFunc(xn)}. This function can be used to create SQL queries where many
table fields have to be checked in the same way for some criterion (see
example).
}
\examples{
## Build SQL query finding records in table t in which at least
## one of the table fields f1 to f100 is NULL.
sql <- sprintf("SELECT * FROM t WHERE \%s", 
               hsSqlExOr(paste("f", 1:100, sep = ""), "isNull"))
  
## Show the SQL string
sql
  
## Output (middle part omitted):
# SELECT * FROM t WHERE (FALSE) OR isNull(f1) OR 
# isNull(f2) OR isNull(f3) OR ... OR isNull(f100)

}
\seealso{
\code{\link{hsSqlExAnd}}
}
