% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_expressions.R
\name{hsSqlExTimeGroup}
\alias{hsSqlExTimeGroup}
\title{SQL Expression: Time Grouping}
\usage{
hsSqlExTimeGroup(tbl, tsField, interval, cond = "TRUE")
}
\arguments{
\item{tbl}{name of the table}

\item{tsField}{name of the table field containing the timestamp}

\item{interval}{specifies the time period to group by ("y": years, "m":
months, "d": days)}

\item{cond}{additional condition in SQL syntax}
}
\value{
Returns SQL code for grouping timestamps by years, months or days
}
\description{
Generates SQL code for grouping timestamps by years, months or days
}
\examples{
## Show SQL query that gets the number of datasets per 
## day ("d") considering the condition "Q > 0"
hsSqlExTimeGroup("myTable", "myTimestamp", "d", "Q > 0")
  
## Output (reformatted):  
## SELECT DateSerial(Year(hsTS), Month(hsTS), Day(hsTS)) 
##   AS myInterval, Count(*) AS myCount 
## FROM (
##     SELECT myTimestamp AS hsTS FROM myTable WHERE Q > 0
## ) 
## GROUP BY DateSerial(Year(hsTS), Month(hsTS), Day(hsTS)) 
## ORDER BY DateSerial(Year(hsTS), Month(hsTS), Day(hsTS))

}
\seealso{
\code{\link{hsSqlExTimeCond}}
}
