% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_expressions.R
\name{hsSqlExTsFields}
\alias{hsSqlExTsFields}
\title{SQL Expression: Timestamp Info}
\usage{
hsSqlExTsFields(tsField, extraTsFields = 0:11)
}
\arguments{
\item{tsField}{name of the table field containing the timestamp}

\item{extraTsFields}{vector of integers representing different types of
time-stamp information:\cr 0 = the timestamp in POSIXct as it is converted
to by R from the text representation delivered by MS Access, \cr 1 =
Timestamp as character string, \cr 2 = Date, \cr 3 = number of days since
1899-12-30, \cr 4 = number of seconds since midnight of Date \cr 5 = number
of minutes since midnight of Date \cr 6:8 = year, month, day \cr 9:11 =
hours, minutes, seconds \cr}
}
\value{
Returns SQL code for selecting different types of information on 
  the time stamp.
}
\description{
Generates SQL code for selecting different parts of the timestamp
}
\examples{
hsSqlExTsFields("myTimestamp", c(6:11))
  
## Output (re-formatted):
# "Year(myTimestamp) AS myTimestamp_year, 
#   Month(myTimestamp) AS myTimestamp_month, 
#   Day(myTimestamp) AS myTimestamp_day, 
#   Hour(myTimestamp) AS myTimestamp_h, 
#   Minute(myTimestamp) AS myTimestamp_m, 
#   Second(myTimestamp) AS myTimestamp_s"

}
\seealso{
\code{\link{hsGetTimeSeries}}
}
