% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsTables.R
\name{hsTables}
\alias{hsTables}
\title{Available tables in database}
\usage{
hsTables(
  mdb,
  excludeSystemTables = grepl("\\\\.(mdb|accdb)$", mdb),
  namesOnly = TRUE,
  use2007Driver = NULL,
  dbg = FALSE
)
}
\arguments{
\item{mdb}{full path to MS Access database file (extension \dQuote{.mdb} or 
\dQuote{.accdb}) or MS Excel file (extension \dQuote{.xls} or 
\dQuote{.xlsx}).}

\item{excludeSystemTables}{if TRUE (default), system tables are excluded from
the table list, else included.}

\item{namesOnly}{if TRUE, only table names are returned. Default: TRUE}

\item{use2007Driver}{if TRUE the functions odbcConnectAccess2007 and
odbcConnectExcel2007 are used instead of odbcConnectAccess and
odbcConnectExcel, respectively}

\item{dbg}{if TRUE, debug messages are shown}
}
\value{
data.frame with columns \emph{TABLE_CAT}, \emph{TABLE_SCHEM}, 
  \emph{TABLE_NAME}, \emph{TABLE_TYPE}, \emph{REMARKS}, see sqlTables of 
  RODBC package.
}
\description{
Returns a data.frame as returned by sqlTables, containing information on the 
tables contained in the database.  Opening of the database connection, 
getting the list of tables and closing of the database connection is done 
within this function.
}
\examples{
\dontrun{
## Get names of tables in the example database
## (only on Windows)

if (.Platform$OS.type == "windows") {

  tnames <- hsTables(xmdb(), namesOnly = TRUE)
  
  ## Exclude system tables by filtering for table names
  ## not starting with '^MSys'
  tNonSys <- grep("^MSys", tnames, invert = TRUE, value = TRUE)
  
  ## Print the names of the non-system tables.
  cat(paste(tNonSys, "\n"))
}
   
## Ouput:
# tbl_Hyd
#  tbl_Qua
#  ...
}

}
\seealso{
\code{\link{hsFields}}
}
