% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{isExcel2003File}
\alias{isExcel2003File}
\title{Is this an XLS file?}
\usage{
isExcel2003File(filepath)
}
\arguments{
\item{filepath}{(vector of) path(s) to the file(s) to be checked for .xls
extension}
}
\value{
(vector of) logical.
}
\description{
Is this an XLS file?
}
