% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{isExcel2007File}
\alias{isExcel2007File}
\title{Is this an XLSX file?}
\usage{
isExcel2007File(filepath)
}
\arguments{
\item{filepath}{(vector of) path(s) to the file(s) to be checked for .xlsx
extension}
}
\value{
(vector of) logical.
}
\description{
Is this an XLSX file?
}
