% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{isExcelFile}
\alias{isExcelFile}
\title{Is this an Excel file?}
\usage{
isExcelFile(filepath)
}
\arguments{
\item{filepath}{(vector of) path(s) to the file(s) to be checked for .xls
or .xlsx extension}
}
\value{
(vector of) logical.
}
\description{
Is this an Excel file?
}
