% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{isMySQL}
\alias{isMySQL}
\title{Is the Given Database of Type MySQL?}
\usage{
isMySQL(db, ...)
}
\arguments{
\item{db}{database file (*.mdb, *.accdb, *.xls, *.xlsx) or name of ODBC
database}

\item{\dots}{arguments passed to \code{\link{hsOpenDb}}, e.g.
\emph{use2007Driver}}
}
\value{
TRUE if \emph{db} is a MySQL database, else FALSE
}
\description{
Is the Given Database of Type MySQL?
}
