% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{keyValuesToSql}
\alias{keyValuesToSql}
\title{List of Key = Value Pairs to SQL Expressions}
\usage{
keyValuesToSql(keyValues, filter, like = filter)
}
\arguments{
\item{keyValues}{list of \code{key = value} pairs}

\item{filter}{logical. If \code{TRUE} the target is an SQL filter expression,
otherwise an SQL SET expression.}

\item{like}{passed to \code{\link{keyValueToSql}}}
}
\description{
List of Key = Value Pairs to SQL Expressions
}
\examples{
keyValues <- list(name = "Peter", birth = as.POSIXct("1999-09-09"))

setCurrentSqlDialect("msaccess")
cat(keyValuesToSql(keyValues, filter = TRUE))
cat(keyValuesToSql(keyValues, filter = TRUE, like = FALSE))
cat(keyValuesToSql(keyValues, filter = FALSE))

setCurrentSqlDialect("mysql")
cat(keyValuesToSql(keyValues, filter = TRUE))
cat(keyValuesToSql(keyValues, filter = TRUE, like = FALSE))
cat(keyValuesToSql(keyValues, filter = FALSE))

}
