% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{keyValuesToSqlAssignment}
\alias{keyValuesToSqlAssignment}
\title{List of Key = Value Pairs to SQL Assignment}
\usage{
keyValuesToSqlAssignment(keyValues)
}
\arguments{
\item{keyValues}{list of \code{key = value} pairs}
}
\description{
List of Key = Value Pairs to SQL Assignment
}
