% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{keyValuesToSqlAssignment2}
\alias{keyValuesToSqlAssignment2}
\title{Key Values to SQL Assignment (2)}
\usage{
keyValuesToSqlAssignment2(keyvalues)
}
\arguments{
\item{keyvalues}{list of \code{key = value pairs}}
}
\value{
list with elements \emph{fieldList} and \emph{valueList}
}
\description{
Key Values to SQL Assignment (2)
}
