% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{keyValuesToSqlFilter}
\alias{keyValuesToSqlFilter}
\title{List of Key = Value Pairs to SQL Filter Expression}
\usage{
keyValuesToSqlFilter(keyValues, like = FALSE)
}
\arguments{
\item{keyValues}{list of \code{key = value} pairs}

\item{like}{if \code{TRUE}, the SQL comparison operator will be \code{LIKE} 
instead of \code{=}.}
}
\description{
List of Key = Value Pairs to SQL Filter Expression
}
