% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{keyValueToSql}
\alias{keyValueToSql}
\title{Generate SQL Filter or Assignment Expression}
\usage{
keyValueToSql(cname, cvalue, like = TRUE, filter = TRUE)
}
\arguments{
\item{cname}{field name}

\item{cvalue}{field value}

\item{like}{if \code{TRUE}, the SQL comparison operator will be \code{LIKE} 
instead of \code{=}.}

\item{filter}{if \code{TRUE} an SQL filter expression is returned, otherwise 
an SQL assignment expression}
}
\value{
(vector of) character representing an SQL expression
}
\description{
Generate SQL Filter or Assignment Expression
}
\examples{
cat(kwb.db:::keyValueToSql("age", 1))
cat(kwb.db:::keyValueToSql("name", "peter"))
cat(kwb.db:::keyValueToSql("name", "peter", like = FALSE))

}
