% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{listValuesToSql}
\alias{listValuesToSql}
\title{Create SQL Tuples from a List}
\usage{
listValuesToSql(x)
}
\arguments{
\item{x}{a list defining \code{key = value} pairs}
}
\value{
vector of character strings each of which represents one assignment
  in \code{x}
}
\description{
Create SQL Tuples from a List
}
\examples{
x <- list(name = "Peter", birthday = as.POSIXct("1981-12-13"))

setCurrentSqlDialect("msaccess")
cat(listValuesToSql(x))

# Note that the representation of a date and time is different in MySQL
setCurrentSqlDialect("mysql")
cat(listValuesToSql(x))

}
