% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lookupRecord}
\alias{lookupRecord}
\title{Lookup Record}
\usage{
lookupRecord(db, tableName, keyAssignments, idField = hsFields(db,
  tableName)[1], dbg = FALSE, use2007Driver = NULL)
}
\arguments{
\item{db}{full path to MS Access/Excel database or name of ODBC data source}

\item{tableName}{name of table in which record is to be looked up}

\item{keyAssignments}{key-value-assignments used to identify the record to be looked up.
The assignments are defined in forms of a list, e.g. list(key1 = "value1",
key2 = "value2").}

\item{idField}{name of ID field, default: name of first table field}

\item{dbg}{if TRUE, debug messages are shown}

\item{use2007Driver}{passed to \code{\link{isMySQL}}}
}
\description{
Looks up a record in a database table and returns the ID of the record or
  NULL if the record does not exist.
}
