% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_relations.R
\name{merge_relations}
\alias{merge_relations}
\title{Merge Table Schema with Relationship Information}
\usage{
merge_relations(schemata, relations)
}
\arguments{
\item{schemata}{list as returned by \code{\link{read_schemata}}}

\item{relations}{data frame as returned by \code{\link{read_relations}}}
}
\description{
Merge Table Schema with Relationship Information
}
