% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{openAdequateConnectionOrStop}
\alias{openAdequateConnectionOrStop}
\title{Open Adequate Connection or Stop}
\usage{
openAdequateConnectionOrStop(db, use2007Driver = NULL, dbg = FALSE,
  DBMSencoding = "", ...)
}
\arguments{
\item{db}{database name or file}

\item{use2007Driver}{if TRUE the functions odbcConnectAccess2007 and
odbcConnectExcel2007 are used instead of odbcConnectAccess and
odbcConnectExcel, respectively}

\item{dbg}{if \code{TRUE}, debug messages are shown}

\item{DBMSencoding}{finally passed to \code{odbcDriverConnect}. Default: "",
You may want to use: "UTF-8"}

\item{\dots}{further arguments passed to \code{odbcConnectionAccess}, 
\code{odbcConnectionExcel} or \code{\link[RODBC]{odbcConnect}}}
}
\description{
Open Adequate Connection or Stop
}
