% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{printDatabaseSchema}
\alias{printDatabaseSchema}
\title{Print Database Schema}
\usage{
printDatabaseSchema(dbSchema)
}
\arguments{
\item{dbSchema}{database schema as returned by
\code{\link{getDatabaseSchema}}}
}
\description{
Print Database Schema
}
