% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_R_interface.R
\name{rcode_DatabaseInterface}
\alias{rcode_DatabaseInterface}
\title{R-code for Functions to Access Database Tables}
\usage{
rcode_DatabaseInterface(db = NULL, dbSchema = NULL, tableNamePattern = "",
  functionPrefix = "db_", pattern = "^", dbg = TRUE, sql.dbg = TRUE,
  create.create.functions = FALSE)
}
\arguments{
\item{db}{database name or file}

\item{dbSchema}{database schema as returned by
\code{\link{getDatabaseSchema}}}

\item{tableNamePattern}{pattern matching the names of tables/views for which
an accessor function is to be generated}

\item{functionPrefix}{prefix to be given to each generated function. Default:
"db_"}

\item{pattern}{pattern matching the part of the table name that will be
replaced with \emph{functionPrefix}. Default: "^", i.e. the prefix will be
appended to the start of the table name}

\item{dbg}{if TRUE, progress messages are shown during the creation of the
interface functions}

\item{sql.dbg}{if TRUE, SQL strings used to access the database are shown
when calling the interface functions}

\item{create.create.functions}{if TRUE, functions for creating new records in
the database tables are created}
}
\description{
R-code for Functions to Access Database Tables
}
