% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_relations.R
\name{read_relations}
\alias{read_relations}
\title{Read Relations from MSysRelationships Table CSV Export}
\usage{
read_relations(file)
}
\arguments{
\item{file}{path to "MSysRelationships.csv" as exported from MS Access by
right clicking on the (hidden) MSysRelationships table and selecting
"Export->Text file->OK"}
}
\value{
data frame with columns \code{table}, \code{field}, \code{refTable}, 
  \code{refField}
}
\description{
Read Relations from MSysRelationships Table CSV Export
}
