% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{readSqlCommandsFromFile}
\alias{readSqlCommandsFromFile}
\title{Read SQL Commands from File}
\usage{
readSqlCommandsFromFile(sqlScript)
}
\arguments{
\item{sqlScript}{full path to file containing SQL commands}
}
\description{
Lines starting with "--" or "#" are ignored. SQL commands must be separated
by semicolon and end of line character (\\n).
}
