% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{renamesToFieldList}
\alias{renamesToFieldList}
\title{List of Renamings to Field Selection String}
\usage{
renamesToFieldList(renames, source.in.brackets = TRUE,
  target.in.brackets = TRUE)
}
\arguments{
\item{renames}{list of key = value pairs defining renamings from the keys to
the values}

\item{source.in.brackets}{if \code{TRUE} (default) the source expressions
will be enclosed in square brackets: \code{[source] AS <target>}}

\item{target.in.brackets}{if \code{TRUE} (default) the target names will be
enclosed in square brackets: \code{<source> AS [target]}}
}
\description{
Convert a list of renamings to a field selection string that can be used as
\code{fields} argument in a call to \code{kwb.db::hsGetTable}
}
