% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{renamesToQueries}
\alias{renamesToQueries}
\title{List of Renamings to SQL Queries}
\usage{
renamesToQueries(renamesList)
}
\arguments{
\item{renamesList}{list of renaming lists. A renaming list is a list of 
\code{key = value} pairs defining renamings from the keys to the values.}
}
\value{
list of character each of which represents an SQL query
}
\description{
Convert a list of renamings to a list of SQL queries each of which can be
used to select and rename the given fields from a database.
}
