% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runSqlCommandsFromFile}
\alias{runSqlCommandsFromFile}
\title{Run SQL Commands from File}
\usage{
runSqlCommandsFromFile(db, sqlFile, ...)
}
\arguments{
\item{db}{Full path to MS Access mdb file or ODBC database name}

\item{sqlFile}{full path to file containing Ms Access SQL commands}

\item{\dots}{further arguments passed to \code{\link{hsSqlQuery}}}
}
\description{
Run SQL Commands from File
}
