% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{safeTableName}
\alias{safeTableName}
\title{Safe Table Name}
\usage{
safeTableName(tableName, sqlDialect = getCurrentSqlDialect(warn = FALSE))
}
\arguments{
\item{tableName}{name of table to be quoted}

\item{sqlDialect}{one of \code{c("mysql", "msaccess")}}
}
\description{
MS Access: table name enclosed in brackets "[" and "]", else: table name
enclosed in backquotes "`"
}
