% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_R_interface.R
\name{selectFromTable}
\alias{selectFromTable}
\title{Select from Table}
\usage{
selectFromTable(db, tableName, arguments, run = TRUE, ...)
}
\arguments{
\item{db}{database name or file}

\item{tableName}{name of table from which to read data}

\item{arguments}{list with elements \code{select}, \code{where}, 
\code{orderBy} and further elements starting with  \code{select_}, 
\code{where_} and  \code{orderBy_}, respectively}

\item{run}{if TRUE (default) the SQL SELECT statement is run otherwise
returned as character string}

\item{\dots}{further arguments passed to \code{\link{hsSqlQuery}}}
}
\description{
Select from Table
}
