% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current.R
\name{setCurrentDb}
\alias{setCurrentDb}
\title{Set Current Database}
\usage{
setCurrentDb(db)
}
\arguments{
\item{db}{full path to MS Access database or ODBC database name}
}
\description{
Set Current Database
}
