% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current.R
\name{setCurrentSqlDialect}
\alias{setCurrentSqlDialect}
\title{Set Current SQL Dialect}
\usage{
setCurrentSqlDialect(dialectName)
}
\arguments{
\item{dialectName}{one of "msaccess", "mysql"}
}
\description{
Set Current SQL Dialect
}
