% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_top_level.R
\name{sqlForInsert}
\alias{sqlForInsert}
\title{Generate SQL INSERT statement}
\usage{
sqlForInsert(tablename, fields, sqlSource,
  sourceAreValues = !grepl(sqlSource, "^SELECT"), ignore = FALSE)
}
\arguments{
\item{tablename}{table name}

\item{fields}{field names, separated by comma}

\item{sqlSource}{value tupels of form (value1.1, value1.2, value1.3, ...)
(value2.1, value2.2, value2.3, ...) ... or SQL SELECT statement providing
these tupels}

\item{sourceAreValues}{if TRUE, \emph{sqlSource} is expected to be an SQL
query providing data to be inserted -> no keyword VALUES in generated SQL
code}

\item{ignore}{if TRUE the keyword IGNORE is inserted between INSERT and INTO
in the SQL statement -> no error will be given if data to insert already
exists}
}
\description{
Generate SQL INSERT statement
}
