% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_top_level.R
\name{sqlForInsertDataFrame}
\alias{sqlForInsertDataFrame}
\title{Generate SQL INSERT Statement}
\usage{
sqlForInsertDataFrame(tablename, dataFrame, ignore = FALSE, ...)
}
\arguments{
\item{tablename}{table name}

\item{dataFrame}{data frame with column names representing table field names}

\item{ignore}{if TRUE the keyword IGNORE is inserted between INSERT and INTO
in the SQL statement -> no error will be given if data to insert already
exists}

\item{\dots}{further arguments passed to \code{\link{sqlForInsert}}}
}
\description{
Generate SQL INSERT statement to insert values in a data frame
}
