% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_top_level.R
\name{sqlForInsertFromSelect}
\alias{sqlForInsertFromSelect}
\title{SQL for INSERT FROM SELECT}
\usage{
sqlForInsertFromSelect(target.table, source.table, fields)
}
\arguments{
\item{target.table}{name of target table}

\item{source.table}{name of source table or SQL providing source data}

\item{fields}{vector of character with field names}
}
\description{
Generate SQL INSERT statement of the form INSERT INTO target.table (fields)
  SELECT fields FROM source.table
}
