% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_top_level.R
\name{sqlForInsertIgnoreInMsAccess}
\alias{sqlForInsertIgnoreInMsAccess}
\title{SQL for "INSERT IGNORE" in MS Access}
\usage{
sqlForInsertIgnoreInMsAccess(db, table.source, table.target,
  uniqueFields = NA)
}
\arguments{
\item{db}{database name or file}

\item{table.source}{name of source table}

\item{table.target}{name of target table}

\item{uniqueFields}{names of unique fields}
}
\description{
Returns SQL for inserting all records from table.source that are not yet
  contained in table.target into table.target
}
