% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_top_level.R
\name{sqlForSelect}
\alias{sqlForSelect}
\title{Generate SQL SELECT statement#'}
\usage{
sqlForSelect(tablename, fields = "*", whereClause = "TRUE", groupBy = "",
  orderBy = "", sqlDialect = getCurrentSqlDialect())
}
\arguments{
\item{tablename}{table name}

\item{fields}{expression to select fields; field names are separated by comma
and alias names may be used, just as SQL accepts, e.g.: "tstamp as
myDateTime, parVal as myValue"; Default: "*"}

\item{whereClause}{where condition; Default: "TRUE"}

\item{groupBy}{GROUP BY-clause, Default: "" (no grouping)}

\item{orderBy}{ORDER BY-clause, Default: "" (no sorting of results)}

\item{sqlDialect}{one of \code{c("mysql", "msaccess")}}
}
\description{
Generate SQL SELECT statement#'
}
