% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_top_level.R
\name{sqlForSelectByKey}
\alias{sqlForSelectByKey}
\title{Generate SQL SELECT Statement}
\usage{
sqlForSelectByKey(tablename, fields = "*", keyValues = NULL)
}
\arguments{
\item{tablename}{table name}

\item{fields}{expression to select fields; field names are separated by comma
and alias names may be used, just as SQL accepts, e.g.: "tstamp as
myDateTime, parVal as myValue"; Default: "*"}

\item{keyValues}{list of "key=value" pairs with the keys being valid field
names of \emph{table}}
}
\description{
Generate SQL SELECT statement (key field values instead of where clause)
}
