% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_top_level.R
\name{sqlForUpdate}
\alias{sqlForUpdate}
\title{Generate SQL UPDATE Statement}
\usage{
sqlForUpdate(tablename, keyValues, whereClause, ignore = FALSE)
}
\arguments{
\item{tablename}{table name}

\item{keyValues}{assignments as list of \emph{key=value} pairs with the keys
representing valid fields of table \emph{tablename}}

\item{whereClause}{where condition}

\item{ignore}{if TRUE the keyword IGNORE is inserted between UPDATE and INTO
in the SQL statement -> no error will be given if updating fails, e.g.
because of key constraints}
}
\description{
Generate SQL UPDATE Statement
}
