% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_join.R
\name{sqlFullLeftJoin}
\alias{sqlFullLeftJoin}
\title{Merge SQL Queries to One Query That Performs a Full Left Join}
\usage{
sqlFullLeftJoin(sqls, key)
}
\arguments{
\item{sqls}{list of SQL queries each of which is expected to contain an
attribute \code{alias} giving the alias name for the query}

\item{key}{name of the primary key field, being selected in each SQL query
contained in \code{sqls}}
}
\value{
vector of character of length one representing the result of "left
  join"-ing all sub-queries given in \code{sqls}
}
\description{
Merge SQL Queries to One Query That Performs a Full Left Join
}
\examples{
sql <- sqlFullLeftJoin(key = "id", list(
  structure("SELECT id, field_1 from table_1", alias = "t1"),
  structure("SELECT id, field_2, field_3 from table_2", alias = "t2"),
  structure("SELECT id, field_4 from table_3", alias = "t3")
))

cat(sql)

}
