% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_join.R
\name{sqlJoinExpression}
\alias{sqlJoinExpression}
\title{Create an SQL JOIN Expression}
\usage{
sqlJoinExpression(left, right, condition, type = "INNER")
}
\arguments{
\item{left}{left part of JOIN (e.g. table name)}

\item{right}{right part of JOIN (e.g. table name)}

\item{condition}{condition}

\item{type}{one of c("LEFT", "RIGHT", "INNER")}
}
\description{
Create an SQL JOIN Expression
}
