% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_join.R
\name{sqlLeftJoin}
\alias{sqlLeftJoin}
\title{Generate SQL for LEFT JOIN}
\usage{
sqlLeftJoin(sqlSource, tablesAndIDs, fields = "*")
}
\arguments{
\item{sqlSource}{SQL of subquery that provides the "base" table on the left}

\item{tablesAndIDs}{named character vector with the names representing the
names of the tables to be joined and the values representing the ID fields
of these tables}

\item{fields}{fields to be selected}
}
\description{
Generate SQL for LEFT JOIN
}
