% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_join.R
\name{sqlLeftJoinBody}
\alias{sqlLeftJoinBody}
\title{Generate (Base Part of) SQL for LEFT JOIN}
\usage{
sqlLeftJoinBody(
  leftSql,
  rightTable,
  id,
  idLeft = id,
  useAlias = (id != idLeft),
  aliasName = "tbase"
)
}
\arguments{
\item{leftSql}{SQL of subquery that provides the "base" table on the left}

\item{rightTable}{name of "right" table}

\item{id}{name of ID field of "right" table (must correspond to a field
returned by \emph{leftSql})}

\item{idLeft}{name of ID field of "left" table}

\item{useAlias}{if TRUE, the alias given in \emph{aliasName} is given to the
subquery \emph{leftSql}. Default: FALSE}

\item{aliasName}{alias name to be used if \emph{useAlias} is TRUE. Default:
"tbase"}
}
\description{
Generate (Base Part of) SQL for LEFT JOIN
}
