% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_join.R
\name{sqlLeftJoinExpression}
\alias{sqlLeftJoinExpression}
\title{Create an SQL LEFT JOIN Expression}
\usage{
sqlLeftJoinExpression(left, right, condition)
}
\arguments{
\item{left}{left part of JOIN (e.g. table name)}

\item{right}{right part of JOIN (e.g. table name)}

\item{condition}{condition}
}
\description{
Create an SQL LEFT JOIN Expression
}
