% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{getDatabaseSchema}
\alias{getDatabaseSchema}
\title{Get Database Schema}
\usage{
getDatabaseSchema(db)
}
\arguments{
\item{db}{full path to database (*.mdb, *.xls) or name of ODBC database}
}
\value{
list with elements \emph{tables} and \emph{relationships}. Element
  \emph{tables} is a list o named elements with the name representing the
  table names and the elements being lists describing the table...
}
\description{
Get Database Schema
}
