% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hsDumpMdb}
\alias{hsDumpMdb}
\title{Export Database Tables to CSV Files}
\usage{
hsDumpMdb(mdb, ptrn = "^tbl", tdir = file.path(dirname(mdb), gsub("\\\\.",
  "_", basename(mdb))))
}
\arguments{
\item{mdb}{full path to database}

\item{ptrn}{pattern matching names of tables to be exported. Default: "^tbl",
i.e. tables starting with "tbl"}

\item{tdir}{target directory. By default a new directory is created in the
same directory as mdb resides in. The new directory has the same name as
the database file with dots substituted with underscores}
}
\description{
Exports all tables of a database of which the names match a given pattern to
csv files.
}
